Be more readable

+++ install.sh
@@ -14,26 +14,22 @@ readonly RED="${LC}31${RC}"
 
 rename_or_remove_if_exist() {
   [[ ! -e "$1" ]] && return 0
+  [[   -h "$1" ]] && { rm "$1"; return 0; }
 
-  if [[ -h "$1" ]]; then
-    rm "$1"
-  else
-    rename_or_remove_if_exist "${1}.bak"
+  rename_or_remove_if_exist "${1}.bak"
 
-    echo -n 'backup) '
-    mv -v "$1" "${1}.bak"
-  fi
+  echo -n 'backup) '
+  mv -v "$1" "${1}.bak"
 }
 
 printf "${BOLD}%s${NONE}\n" 'Creating symbolic links...'
 
-for fpath in "${VIRTUAL_HOME}"/.??*
-do
+for fpath in "${VIRTUAL_HOME}"/.??*; do
   fname=$(basename "${fpath}")
 
   if [[ "${fname}" == '.??*' ]]; then
-    { printf "${BOLD}${RED}%s${NONE}" 'Error: '
-      echo "No dotfiles in ~/${VIRTUAL_HOME}."; } 1>&2
+    printf 1>&2 "${BOLD}${RED}%s: ${NONE}%s\n" \
+      'Error' "No dotfiles in ~/${VIRTUAL_HOME}."
     exit 1
   fi
